int fixes

+++ virtualgrid.py
@@ -242,7 +242,7 @@ class PhysicalGridWrapper_2(monome.GridWrapper):
 class myGridBuffer:
     def __init__(self, width, height):
         #self.levels = [[0 for col in range(width)] for row in range(height)]
-        self.levels = np.zeros((width,height), dtype=int)
+        self.levels = np.zeros((width,height), dtype=intc)
         self.width = width
         self.height = height
